Build: Better defaults, clearer npm commands, make
The makefile is now the primary build tool and developpement
environement tool. It makes the bridge between the docker container
management the building of the front-end and backend and also runs the
formating.

The pom.xml now has better defaults. By default it will build the java
and the front-end. Disable the `webapp` profile to not build the
front-end.

Npm commands are clearer `npm run fmt` will not formate the backend
anymore and just run prettier on the front-end.