Specify document units routes + initial gen config
Created the `openapi.yaml` file in ressources. It defines the api spec
in yaml and covers the routes under `/doc_units`.
Also included a `config.yaml` file that add an initial set of
configuration options for the `spring` generator from the
`open-api-generator` project.

In the future, when the implementation of the api will start, adding the
 generator to the build, in the `pom.xml`, will be required.
For now a `test-gen.sh` script can be used to inspect the result of the
code generation. The resulting generated code will be put in the
`test-gen` folder in the root of the project.

The code generation is configured to create an Interface by tag that
will contain all the routes. (eg a `DocUnitsApi`).