Make NumaHOP output logs in json.
This patch allows NumaHOP to be able to output structured logs in json
format.

The functionality that allows the user to download log files was also
removed as it was only accessible from the admin account. We assume the
admin account has access to the machine and so this functionality was
redundant and not really useful.

Structured logs are disable by default in the configuration. And the
default configuration is made to match the previous comportment.

This configuration format is not definitive as structured logs are
available in a future version of Spring-Boot.