Automatic refactoring for @Query annotations.
When the formatter was changed to spring-javaformat the queries were
written with concatenated strings emulating multi-line strings. But
spring-javaformat is stricter and the readability was worsened.

The refactor used on each file is contained inside the refactor.sh
script.

This refactor was made using vim/neovim commands. On each file the following
commands where run in order:
- `%s/"\(\s\|\n\)*+\s"//g`  This merges the separated strings.
- `%s/Query("\([^"]*\)")/Query("""\r\t\t\1\r\t\t""")/` This replace the
  merged string by a block string.
- `g/^\s*from/norm ^wwyw^PIselect ` This fixes the implicit select
  from queries.
- `g/\c^\s*\(select\|delete\|update\)/s/\c\(inner join\|left join\|join\|where\|order by\)/\r\t\t\1/g`
This puts one statement per line.

This change aims to improve the readability of queries.
As an added bonus some code editors are able to provide syntax
highlighting to the queries when they are contained in one string.